chore: nouvel espace de nom pour C++
