+++ cpp/include/Layout.hpp
@@ -1,7 +1,12 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
 #ifndef LAYOUT_HPP
 #define LAYOUT_HPP
 
 #include <Adafruit_SSD1306.h>
+#include <Adafruit_GFX.h>
 
 class Layout
 {
