+++ cpp/include/View/ErrorView.hpp
@@ -1,0 +1,14 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#ifndef ERRORVIEW_HPP
+#define ERRORVIEW_HPP
+
+class ErrorView
+{
+public: // pages
+  const void display_error() const;
+};
+
+#endif
