+++ cpp/include/View/HomeView.hpp
@@ -1,0 +1,14 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#ifndef HOMEVIEW_HPP
+#define HOMEVIEW_HPP
+
+class HomeView
+{
+public: // pages
+  const void display_home() const;
+};
+
+#endif
