+++ cpp/src/View/HomeView.cpp
@@ -1,5 +1,8 @@
-#include "View/HomeView.hpp"
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+#include "View/HomeView.hpp"
 #include "Encoder/RotaryEncoder.hpp"
 #include "Layout.hpp"
 #include "View.hpp"
