+++ cpp/src/main.cpp
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
 #include <Adafruit_SSD1306.h>
 #include <ArduinoJson.h>
 #include <SPI.h>
