--- include/View/ErrorView.hpp
@@ -1,10 +1,0 @@
-#ifndef ERRORVIEW_HPP
-#define ERRORVIEW_HPP
-
-class ErrorView
-{
-public: // pages
-  const void display_error() const;
-};
-
-#endif
