--- include/View/HomeView.hpp
@@ -1,10 +1,0 @@
-#ifndef HOMEVIEW_HPP
-#define HOMEVIEW_HPP
-
-class HomeView
-{
-public: // pages
-  const void display_home() const;
-};
-
-#endif
