+++ platformio.ini
@@ -8,12 +8,19 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+[platformio]
+workspace_dir = target/pio
+include_dir = cpp/include
+lib_dir = cpp/lib
+src_dir = cpp/src
+test_dir = cpp/test
+
 [env:megaatmega2560]
 platform = atmelavr
 board = megaatmega2560
 framework = arduino
-lib_deps =
-	adafruit/Adafruit BusIO@^1.9.5
+lib_deps = 
+	adafruit/Adafruit BusIO@^1.9.6
 	adafruit/Adafruit GFX Library@^1.10.12
 	adafruit/Adafruit SSD1306@^2.5.0
 	adafruit/SD@0.0.0-alpha+sha.041f788250
